+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.3.1 - 2025-01-18
+
+### Fixed
+- The ```thrownException``` member of Context has been changed to a ```Throwable``` so that ```Error```s can also be declared
+  as expected in a test.
+- When an exception is declared via ```expectingException``` but no exception was thrown during the test, the test was passing.
+
 ## 1.3.0 - 2024-12-27
 
 ### Added